refactor: 파일 분리
- index 파일의 함수를 render, utils, index로 세분
