+++ public/utils.js
@@ -1,0 +1,9 @@
+const $ = (selector) => {
+  return document.querySelector(selector);
+};
+
+const range = (n) => {
+  return [...Array(n)];
+};
+
+export { $, range };
